Move UpgradePlan component to separate folder
